+++ phtree/phtree_d_test_filter.cc
@@ -240,6 +240,7 @@ struct CallbackConst {
     }
 };
 
+[[maybe_unused]]
 static void print_id_counters() {
     std::cout << "dc=" << f_default_construct_ << " c=" << f_construct_
               << " cc=" << f_copy_construct_ << " mc=" << f_move_construct_
@@ -407,4 +408,74 @@ TEST(PhTreeTest, TestFilterAPI_KNN) {
     // rvalue
     ASSERT_EQ(treeC.begin_knn_query(3, {2, 3, 4}, DistanceCount<3>{}, FilterConst<3, Id>())->_i, 1);
     f_reset_id_counters();
+}
+
+template <dimension_t DIM>
+double distance(const TestPoint<DIM>& p1, const TestPoint<DIM>& p2) {
+    double sum2 = 0;
+    for (dimension_t i = 0; i < DIM; ++i) {
+        double d2 = p1[i] - p2[i];
+        sum2 += d2 * d2;
+    }
+    return sqrt(sum2);
+};
+
+template <dimension_t DIM>
+void referenceSphereQuery(
+    std::vector<TestPoint<DIM>>& points,
+    TestPoint<DIM>& center,
+    double radius,
+    std::set<size_t>& result) {
+    for (size_t i = 0; i < points.size(); i++) {
+        auto& p = points[i];
+        if (distance(center, p) <= radius) {
+            result.insert(i);
+        }
+    }
+}
+
+// We use 'int&' because gtest does not compile with assertions in non-void functions.
+template <dimension_t DIM>
+void testSphereQuery(TestPoint<DIM>& center, double radius, size_t N, int& result) {
+    TestTree<DIM, size_t> tree;
+    std::vector<TestPoint<DIM>> points;
+    populate(tree, points, N);
+
+    std::set<size_t> referenceResult;
+    referenceSphereQuery(points, center, radius, referenceResult);
+
+    result = 0;
+    auto filter = FilterSphere(center, radius, tree.converter());
+    for (auto it = tree.begin(filter); it != tree.end(); it++) {
+        auto& x = *it;
+        ASSERT_GE(x, 0);
+        ASSERT_EQ(referenceResult.count(x), 1);
+        result++;
+    }
+    ASSERT_EQ(referenceResult.size(), result);
+}
+
+TEST(PhTreeDTest, TestSphereQuery0) {
+    const dimension_t dim = 3;
+    TestPoint<dim> p{-10000, -10000, -10000};
+    int n = 0;
+    testSphereQuery<dim>(p, 0.1, 100, n);
+    ASSERT_EQ(0, n);
+}
+
+TEST(PhTreeDTest, TestSphereQueryMany) {
+    const dimension_t dim = 3;
+    TestPoint<dim> p{0, 0, 0};
+    int n = 0;
+    testSphereQuery<dim>(p, 1000, 1000, n);
+    ASSERT_GT(n, 400);
+    ASSERT_LT(n, 800);
+}
+
+TEST(PhTreeDTest, TestSphereQueryAll) {
+    const dimension_t dim = 3;
+    TestPoint<dim> p{0, 0, 0};
+    int n = 0;
+    testSphereQuery<dim>(p, 10000, 1000, n);
+    ASSERT_EQ(1000, n);
 }